+++ logo.py
@@ -2,17 +2,17 @@ from colorama import Fore, Style, init
 from dotenv import load_dotenv
 import os
 
-# 獲取當前腳本所在目錄
+# Get the current script directory
 current_dir = os.path.dirname(os.path.abspath(__file__))
-# 構建.env文件的完整路徑
+# Build the full path to the .env file
 env_path = os.path.join(current_dir, '.env')
 
-# 加載環境變量，指定.env文件路徑
+# Load environment variables, specifying the .env file path
 load_dotenv(env_path)
-# 獲取版本號，如果未找到則使用默認值
+# Get the version number, using the default value if not found
 version = os.getenv('VERSION', '1.0.0')
 
-# 初始化 colorama
+# Initialize colorama
 init()
 
 CURSOR_LOGO = f"""
@@ -27,6 +27,8 @@ CURSOR_LOGO = f"""
                 Pro Version Activator v{version}
 {Fore.GREEN}
                 Author: Pin Studios | yeongpin
+
+      Github: https://github.com/yeongpin/cursor-free-vip
 {Fore.RED}
         Press 5 to change language | 按下 5 键切换语言
 {Style.RESET_ALL}